requirements.txt
streamlit
ultralytics
torch
torchvision
torchaudio
opencv-python-headless
pillow
numpy
pandas
matplotlib
seaborn
tqdm
PyYAML
scikit-learn